Pop field changed from Double to Int

+++ data/src/main/java/es/jnsoft/data/model/HourlyData.kt
@@ -9,7 +9,7 @@ data class HourlyData(
     val icon: String,
     val lat: Double,
     val lon: Double,
-    val pop: Double,
+    val pop: Int,
     val pressure: Int,
     val rain: Double,
     val snow: Double,
